+++ Changelog.md
@@ -18,6 +18,10 @@ WinCopies.Util (2.4.0-preview6)
 - Existing items behavior updates:
 	- The WinCopies.Util.Util.StartProcessNetCore method returns the process that is launched.
 
+- Additions:
+	- (Observable)Stack/Queue/LinkedListCollection classes. Fixes #5.
+	- WinCopies.Util.Util.NotApplicable const.
+
 05/21/2020 2.3.0-preview5
 =========================